requirements.txt
# Веб-фреймворк и API
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Работа с PDF
pdfplumber==0.10.3
camelot-py[base]==0.11.0
pypdf2==3.0.1

# ИИ и NLP
ollama==0.1.4
requests==2.31.0

# База данных
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Очередь задач
celery==5.3.4
redis==5.0.1

# Генерация документов
python-docx==0.8.11
jinja2==3.1.2

# Утилиты
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pandas==2.1.3

# Тестирование
pytest==7.4.3
pytest-asyncio==0.21.1

# Мониторинг
prometheus-client==0.19.0



